+++ tree_shaking/__main__.py
@@ -1,6 +1,7 @@
 import neoprint as np
 from argsense import cli
 from lk_utils import fs
+from neoprint import print
 
 from . import insight
 from .export import dump_tree_from_config_file
@@ -22,28 +23,63 @@ def dump_tree(config_file: str, dir_o: str = '', dry_run: int = 0) -> None:
 
 
 def trace(
-    entry_script: str, target_module: str, search_paths: str = '$venv'
+    entry_script: str,
+    target_module: str,
+    search_paths: str = '$venv;$proj;$entry',
+    concise: bool = False,
 ) -> None:
+    """
+    Args:
+        entry_script (-e):
+        target_module (-t):
+        search_paths (-s):
+        concise (-c):
+    """
+
     entry_script = fs.abspath(entry_script)
+
+    def get_entry_dir():
+        return fs.parent(entry_script)
+
+    def get_proj_dir():
+        temp_dir = fs.parent(entry_script)
+        while True:
+            if fs.exist('{}/pyproject.toml'.format(temp_dir)):
+                return temp_dir
+            else:
+                temp_dir = fs.parent(temp_dir)
+
+    def get_venv_dir():
+        temp_dir = fs.parent(entry_script)
+        while True:
+            if fs.exist('{}/.venv'.format(temp_dir)):
+                assert fs.exist('{}/.venv/Lib/site-packages'.format(temp_dir))
+                return '{}/.venv/Lib/site-packages'.format(temp_dir)
+            else:
+                temp_dir = fs.parent(temp_dir)
+
+    scopes = set()
     for p in search_paths.split(';'):
-        if p == '$venv':
-            temp_dir = fs.parent(entry_script)
-            while True:
-                if fs.exist('{}/.venv'.format(temp_dir)):
-                    assert fs.exist(
-                        '{}/.venv/Lib/site-packages'.format(temp_dir)
-                    )
-                    p = '{}/.venv/Lib/site-packages'.format(temp_dir)
-                    np.print('add venv to scope', p, ':v')
-                    path_scope.add_scope(p)
-                    path_scope.add_scope(temp_dir)
-                    break
-                else:
-                    temp_dir = fs.parent(temp_dir)
+        if p[0] == '$':
+            a, b = p.split('/', 1) if '/' in p else (p, '')
+            c = (
+                get_entry_dir()
+                if a == '$entry'
+                else get_proj_dir()
+                if a == '$proj'
+                else get_venv_dir()
+            )
+            scopes.add(c)
+            if b:
+                scopes.add(c + '/' + b)
         else:
-            path_scope.add_scope(fs.abspath(p))
+            scopes.add(fs.abspath(p))
+    print(sorted(scopes), ':nl')
+    for p in sorted(scopes):
+        path_scope.add_scope(p)
+
     with np.scope('tracing'):
-        insight.trace(entry_script, target_module)
+        insight.trace(entry_script, target_module, concise=concise)
 
 
 cli.add_cmd(build_module_graphs)